+++ src/components/StripeBankAccountTokenInputField/StripeBankAccountTokenInputField.util.js
@@ -6,6 +6,14 @@ import config from '../../config';
 export const ACCOUNT_NUMBER = 'accountNumber';
 // Australian equivalent for routing number
 export const BSB = 'bsb';
+// Needed for creating full routing number in Canada
+export const INSTITUTION_NUMBER = 'institutionNumber';
+// Needed for creating full routing number in Canada
+export const TRANSIT_NUMBER = 'transitNumber';
+// Needed for creating full routing number in Hong Kong
+export const CLEARING_CODE = 'clearingCode';
+// Needed for creating full routing number in Hong Kong
+export const BRANCH_CODE = 'branchCode';
 // International bank account number (e.g. EU countries use this)
 export const IBAN = 'iban';
 // Routing number to separate bank account in different areas
@@ -15,7 +23,17 @@ export const SORT_CODE = 'sortCode';
 
 // Currently supported bank account inputs
 // the order here matters: account number input is asked after routing number and its equivalents
-export const BANK_ACCOUNT_INPUTS = [BSB, SORT_CODE, ROUTING_NUMBER, ACCOUNT_NUMBER, IBAN];
+export const BANK_ACCOUNT_INPUTS = [
+  BSB,
+  TRANSIT_NUMBER,
+  INSTITUTION_NUMBER,
+  CLEARING_CODE,
+  BRANCH_CODE,
+  SORT_CODE,
+  ROUTING_NUMBER,
+  ACCOUNT_NUMBER,
+  IBAN,
+];
 
 export const supportedCountries = config.stripe.supportedCountries.map(c => c.code);
 
@@ -117,7 +135,7 @@ export const translateStripeError = (country, intl, stripeError) => {
  *
  * @return {Object} key - value in Object literal.
  */
-export const mapInputsToStripeAccountKeys = (inputType, value) => {
+export const mapInputsToStripeAccountKeys = (country, values) => {
   // Stripe documentation speaks about actual bank account terms of different countries
   // (like BSB, sort code, routing number), but all of those get mapped to one of
   // the two different request keys: routing_number or account_number
@@ -126,17 +144,63 @@ export const mapInputsToStripeAccountKeys = (inputType, value) => {
   // We use those country specific terms since we want to show correct labels and errors for users,
   // so this mapping is needed before sending data to Stripe.
 
-  switch (inputType) {
-    case IBAN:
-    case ACCOUNT_NUMBER:
-      return { account_number: value };
-    case BSB:
-    case SORT_CODE:
-    case ROUTING_NUMBER:
-      return { routing_number: value };
+  // Stripe fails if there are spaces within the number, this is
+  // why we have to clean value up first.
+
+  switch (country) {
+    case 'AT':
+    case 'BE':
+    case 'DK':
+    case 'FI':
+    case 'FR':
+    case 'DE':
+    case 'IE':
+    case 'IT':
+    case 'LU':
+    case 'NL':
+    case 'PT':
+    case 'ES':
+    case 'SE':
+    case 'CH':
+    case 'NO':
+      return { account_number: cleanedString(values[IBAN]) };
+    case 'NZ':
+      // NZ account number is typically presented in the format xx-xxxx-xxxxxxx-xxx
+      // '-' separators must be removed before sending value to Stripe API
+      return { account_number: cleanedString(values[ACCOUNT_NUMBER]).replace(/-/g, '') };
+    case 'AU':
+      return {
+        routing_number: cleanedString(values[BSB]),
+        account_number: cleanedString(values[ACCOUNT_NUMBER]),
+      };
+    case 'CA':
+      return {
+        routing_number: cleanedString(values[TRANSIT_NUMBER]).concat(
+          cleanedString(values[INSTITUTION_NUMBER])
+        ),
+        account_number: cleanedString(values[ACCOUNT_NUMBER]),
+      };
+    case 'GB':
+      return {
+        routing_number: cleanedString(values[SORT_CODE]),
+        account_number: cleanedString(values[ACCOUNT_NUMBER]),
+      };
+    case 'US':
+      return {
+        routing_number: cleanedString(values[ROUTING_NUMBER]),
+        account_number: cleanedString(values[ACCOUNT_NUMBER]),
+      };
+    case 'HK':
+      return {
+        routing_number: cleanedString(values[CLEARING_CODE]).concat(
+          '-',
+          cleanedString(values[BRANCH_CODE])
+        ),
+        account_number: cleanedString(values[ACCOUNT_NUMBER]),
+      };
 
     default:
-      throw new Error(`Unknown inputType (${inputType}) given to validator`);
+      throw new Error(`Not supported country (${country}) given to validator`);
   }
 };