+++ src/forms/PayoutDetailsForm/PayoutDetailsForm.js
@@ -17,6 +17,7 @@ import {
 import * as validators from '../../util/validators';
 import { isStripeInvalidPostalCode } from '../../util/errors';
 
+import PayoutDetailsAddress from './PayoutDetailsAddress';
 import css from './PayoutDetailsForm.css';
 
 const MIN_STRIPE_ACCOUNT_AGE = 18;
@@ -30,11 +31,6 @@ export const stripeCountryConfigs = countryCode => {
     throw new Error(`Country code not found in Stripe config ${countryCode}`);
   }
   return country;
-};
-
-const requiresAddress = countryCode => {
-  const country = stripeCountryConfigs(countryCode);
-  return country.payoutAddressRequired;
 };
 
 const countryCurrency = countryCode => {
@@ -114,42 +110,53 @@ const PayoutDetailsFormComponent = props => (
         })
       );
 
-      const streetAddressLabel = intl.formatMessage({
-        id: 'PayoutDetailsForm.streetAddressLabel',
-      });
-      const streetAddressPlaceholder = intl.formatMessage({
-        id: 'PayoutDetailsForm.streetAddressPlaceholder',
-      });
-      const streetAddressRequired = validators.required(
-        intl.formatMessage({
-          id: 'PayoutDetailsForm.streetAddressRequired',
-        })
-      );
-
-      const postalCodeLabel = intl.formatMessage({ id: 'PayoutDetailsForm.postalCodeLabel' });
-      const postalCodePlaceholder = intl.formatMessage({
-        id: 'PayoutDetailsForm.postalCodePlaceholder',
-      });
-      const postalCodeRequired = validators.required(
-        intl.formatMessage({
-          id: 'PayoutDetailsForm.postalCodeRequired',
-        })
-      );
-
-      const cityLabel = intl.formatMessage({ id: 'PayoutDetailsForm.cityLabel' });
-      const cityPlaceholder = intl.formatMessage({ id: 'PayoutDetailsForm.cityPlaceholder' });
-      const cityRequired = validators.required(
-        intl.formatMessage({
-          id: 'PayoutDetailsForm.cityRequired',
-        })
-      );
-
-      const showAddressFields = country && requiresAddress(country);
-
       // StripeBankAccountTokenInputField handles the error messages
       // internally, we just have to make sure we require a valid token
       // out of the field. Therefore the empty validation message.
       const bankAccountRequired = validators.required(' ');
+
+      const showPersonalIdNumber =
+        (country && stripeCountryConfigs(country).personalIdNumberRequired) ||
+        (country && stripeCountryConfigs(country).ssnLast4Required);
+
+      const personalIdNumberRequired = validators.required(
+        intl.formatMessage({
+          id: `PayoutDetailsForm.personalIdNumberRequired`,
+        })
+      );
+
+      let personalIdNumberLabel = null;
+      let personalIdNumberPlaceholder = null;
+      let personalIdNumberValid = personalIdNumberRequired;
+
+      if (country === 'US') {
+        personalIdNumberLabel = intl.formatMessage({
+          id: `PayoutDetailsForm.personalIdNumberLabel.US`,
+        });
+        personalIdNumberPlaceholder = intl.formatMessage({
+          id: `PayoutDetailsForm.personalIdNumberPlaceholder.US`,
+        });
+
+        const validSSN = validators.validSsnLast4(
+          intl.formatMessage({
+            id: `PayoutDetailsForm.personalIdNumberValid`,
+          })
+        );
+        personalIdNumberValid = validators.composeValidators(personalIdNumberRequired, validSSN);
+      } else if (country === 'HK') {
+        personalIdNumberLabel = intl.formatMessage({
+          id: `PayoutDetailsForm.personalIdNumberLabel.HK`,
+        });
+        personalIdNumberPlaceholder = intl.formatMessage({
+          id: `PayoutDetailsForm.personalIdNumberPlaceholder.HK`,
+        });
+        const validHKID = validators.validHKID(
+          intl.formatMessage({
+            id: `PayoutDetailsForm.personalIdNumberValid`,
+          })
+        );
+        personalIdNumberValid = validators.composeValidators(personalIdNumberRequired, validHKID);
+      }
 
       const classes = classNames(css.root, className, {
         [css.disabled]: disabled,
@@ -178,7 +185,6 @@ const PayoutDetailsFormComponent = props => (
           <FormattedMessage id="PayoutDetailsForm.stripeConnectedAccountTermsLink" />
         </ExternalLink>
       );
-
       return (
         <Form className={classes} onSubmit={handleSubmit}>
           <div className={css.sectionContainer}>
@@ -245,48 +251,8 @@ const PayoutDetailsFormComponent = props => (
                 </option>
               ))}
             </FieldSelect>
-            {showAddressFields ? (
-              <div>
-                <FieldTextInput
-                  id="streetAddress"
-                  name="streetAddress"
-                  disabled={disabled}
-                  className={css.field}
-                  type="text"
-                  autoComplete="street-address"
-                  label={streetAddressLabel}
-                  placeholder={streetAddressPlaceholder}
-                  validate={streetAddressRequired}
-                  onUnmount={() => form.change('streetAddress', undefined)}
-                />
-                <div className={css.formRow}>
-                  <FieldTextInput
-                    id="postalCode"
-                    name="postalCode"
-                    disabled={disabled}
-                    className={css.postalCode}
-                    type="text"
-                    autoComplete="postal-code"
-                    label={postalCodeLabel}
-                    placeholder={postalCodePlaceholder}
-                    validate={postalCodeRequired}
-                    onUnmount={() => form.change('postalCode', undefined)}
-                  />
-                  <FieldTextInput
-                    id="city"
-                    name="city"
-                    disabled={disabled}
-                    className={css.city}
-                    type="text"
-                    autoComplete="address-level2"
-                    label={cityLabel}
-                    placeholder={cityPlaceholder}
-                    validate={cityRequired}
-                    onUnmount={() => form.change('city', undefined)}
-                  />
-                </div>
-              </div>
-            ) : null}
+
+            <PayoutDetailsAddress country={country} intl={intl} disabled={disabled} form={form} />
           </div>
           {country ? (
             <div className={css.sectionContainer}>
@@ -303,7 +269,27 @@ const PayoutDetailsFormComponent = props => (
               />
             </div>
           ) : null}
+
+          {showPersonalIdNumber ? (
+            <div className={css.sectionContainer}>
+              <h3 className={css.subTitle}>
+                <FormattedMessage id="PayoutDetailsForm.personalIdNumberTitle" />
+              </h3>
+              <FieldTextInput
+                id="personalIdNumber"
+                name="personalIdNumber"
+                disabled={disabled}
+                className={css.personalIdNumber}
+                type="text"
+                label={personalIdNumberLabel}
+                placeholder={personalIdNumberPlaceholder}
+                validate={personalIdNumberValid}
+              />
+            </div>
+          ) : null}
+
           {error}
+
           <p className={css.termsText}>
             <FormattedMessage
               id="PayoutDetailsForm.stripeToSText"